refactor: consolidate study actions into useStudy
