+++ src/entities/study-progress/index.ts
@@ -1,2 +1,1 @@
 export { editStudyProgress } from "./api/firestore";
-export type { StudyProgressEdit } from "./model/types";
